requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.0.0
tqdm>=4.62.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
grad-cam>=1.4.0
opencv-python>=4.5.0
matplotlib>=3.5.0
requests>=2.28.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

